Ya registra entradas SIUUUUUUUU (#12)

+++ src/Modelo/Emergencia.java
@@ -6,7 +6,7 @@ public class Emergencia extends Turno{
     private String descripcion;
     private boolean respuestaRapida, ambulatorio, clinica;
     private String vehiculo;
-    private long codigo;
+    private int codigo;
 
     public String getDescripcion() {
         return descripcion;
@@ -48,11 +48,11 @@ public class Emergencia extends Turno{
         this.vehiculo = vehiculo;
     }
 
-    public long getCodigo() {
+    public int getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(long codigo) {
+    public void setCodigo(int codigo) {
         this.codigo = codigo;
     }
